Small updates, divide by zero bug fix.
Trading is still a little crude, but workable and trades relatively flat on loss (in general). Best performer on USDJPY for past year of 90% tick data, but does well in 4/5 currency pairs tested so far. Currently still testing squeeze on pairs using better backtesting data (quantdata managing and dukascopy). Current drawback is memory and speed required to download and convert gigs of ticks to MT4 using the free version of quantdata.